moved more classes to the model module.
